Move db to commons
