+++ commands/emojicloud.py
@@ -9,7 +9,7 @@ from PIL import Image, ImageDraw, ImageFont
 from matplotlib import cm
 from pilmoji import Pilmoji  # Used to generate decent Unicode emoji
 
-import db
+import commons.db as db
 import commons.emoji_cache as emoji_cache
 
 plugin = lightbulb.Plugin("EmojiCloud")
