+++ commands/snark.py
@@ -1,6 +1,6 @@
 import os
 import lightbulb
-import db
+import commons.db as db
 
 plugin = lightbulb.Plugin("Snark")
